+++ GroceryBama/Entities/Statistic.cs
@@ -5,7 +5,7 @@
         // A statistic within 1 month period
         public int GroceryId { get; set; }
         public string StoreName { get; set; }
-        public double TotalProfit { get; set; }
+        public float TotalProfit { get; set; }
         public int TotalItemsSold { get; set; }
     }
 }
